Add main policies and authorization calls

+++ app/controllers/georgia/media_controller.rb
@@ -2,17 +2,20 @@ module Georgia
   class MediaController < ApplicationController
 
     def index
+      authorize Ckeditor::Asset
       redirect_to action: :search
     end
 
     def search
       @asset = Ckeditor::Asset.new
+      authorize Ckeditor::Asset
       search_conditions = Georgia::MediaSearch.new(params).definition
       @search = Ckeditor::Picture.search(search_conditions).page(params[:page])
       @assets = Ckeditor::AssetDecorator.decorate_collection(@search.records)
     end
 
     def create
+      authorize Ckeditor::Asset
       begin
         @assets = params[:assets].map{|asset| CreateMediaAsset.new(asset).call}
         @assets = Ckeditor::AssetDecorator.decorate_collection(@assets)
@@ -23,15 +26,19 @@ module Georgia
     end
 
     def show
+      @asset = Ckeditor::Asset.find(params[:id])
+      authorize @asset
       redirect_to edit_media_path(id: params[:id])
     end
 
     def edit
       @asset = Ckeditor::Asset.find(params[:id])
+      authorize @asset
     end
 
     def update
       @asset = Ckeditor::Asset.find(params[:id])
+      authorize @asset
       if @asset.update_attributes(params[:asset])
         respond_to do |format|
           format.html { redirect_to edit_media_path(@asset), notice: "Asset was successfully updated." }
@@ -48,6 +55,7 @@ module Georgia
     # Destroy multiple assets
     def destroy
       @assets = Ckeditor::Asset.where(id: params[:id])
+      authorize @assets
       # if can?(:destroy, Ckeditor::Asset) and @assets.destroy_all
       if @assets.destroy_all
         respond_to do |format|
@@ -67,24 +75,10 @@ module Georgia
     # We could then have a download spinner while the request is processing, even a progress bar
     def download
       ids = params[:ids].split(',')
-      @files = Ckeditor::Asset.find(ids)
-      t = Tempfile.new("tmp-zip-#{Time.now.to_i}")
-      Zip::OutputStream.open(t.path) do |zos|
-        @files.each do |file|
-          filename = file.filename
-          zos.put_next_entry(filename)
-          tmp_file = Tempfile.new(filename)
-          open(file.url) do |data|
-            tmp_file.write data.read.force_encoding('UTF-8')
-          end
-          zos.print IO.read(tmp_file)
-          tmp_file.close
-        end
-      end
-      filename = "#{Georgia.title.try(:parameterize) || 'georgia'}_assets_#{Time.now.strftime('%Y%m%d%H%M%S')}.zip"
-      t.close
-
-      send_file t.path, type: "application/zip", disposition: 'attachment', filename: filename
+      @files = Ckeditor::Asset.where(id: ids)
+      authorize @files
+      zip_file = Georgia::CompressFiles.new(@files).file
+      send_file zip_file.path, type: "application/zip", disposition: 'attachment', filename: zip_file.filename
     end
 
   end
